Enable the dovecot service persitently, allow override.
The dovecot service is currently started, but not enabled. This
means that after a reboot, the service is not running.

Change this behavior to enable the service by default, which can be
overridden by setting the dovecot:service_persistent pillar to False.

Drive-By: Update the README.rst file to accurately mirror the current
state of the formula, specifically the supported OS variants.